added a note about Scheme versions.
I'm revising this project after several months ignoring it
and finding various shortcomings of the kind that magically
appear in projects over time despite the code not changing.
If you don't believe in magic you're doing coding wrong.
Anyway as a result of all this, I've been making a bunch of
little notes in this README and I probably shouldn't have bothered
checking in until they were all done, but there you are.